requirements.txt
fastapi==0.135.1
google-cloud-aiplatform==1.140.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-pubsub==2.35.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.9.0
google-cloud-videointelligence==2.18.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
torch==2.2.2
torchvision==0.17.2
ultralytics==8.4.21
ultralytics-thop==2.0.18
uvicorn==0.41.0
